Fix raw MFT read byte-count comparison

+++ FileRecovery/MftCandidateScanner.cs
@@ -257,7 +257,7 @@ public sealed class MftCandidateScanner
 
             scanned += usableBytes;
 
-            if (bytesRead < requestBytes)
+            if (bytesRead < (uint)requestBytes)
             {
                 break;
             }
